Feat/dummydata

See merge request hasanro/idatt2106_2022_05_backend!34

+++ src/main/java/com/example/idatt2106_2022_05_backend/controller/UserController.java
@@ -42,6 +42,9 @@ public class UserController {
     @DeleteMapping("/profilePicture")
     @ApiOperation(value = "Endpoint to delete profile picture", response = Response.class)
     public Response deleteProfilePicture(@ModelAttribute UpdatePictureDto updatePictureDto) {
+        if(!securityService.userPicture(updatePictureDto.getId(), updatePictureDto.getUserId())){
+            return new Response("Du har ikke tilgang", HttpStatus.BAD_REQUEST);
+        }
         try {
             return userService.deleteProfilePicture(updatePictureDto.getUserId(),
                     updatePictureDto.getMultipartFile().getBytes());
